% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRchi2treeQSM.R
\docType{methods}
\name{aRchi2treeQSM}
\alias{aRchi2treeQSM}
\alias{aRchi2treeQSM,aRchi-method}
\title{Generate a QSM in treeQSM format}
\usage{
aRchi2treeQSM(aRchi)

\S4method{aRchi2treeQSM}{aRchi}(aRchi)
}
\arguments{
\item{aRchi}{the object of class aRchi to write}
}
\description{
a data.table with treeQSM format (v2.3.3)
}
