% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_aRchi.R
\name{build_aRchi}
\alias{build_aRchi}
\title{Read an aRchi file}
\usage{
build_aRchi(QSM, point_cloud, keep_original = FALSE)
}
\arguments{
\item{QSM}{A data.table obtained from [read_QSM()] function}

\item{point_cloud}{A point cloud. Either a LAS or a data.table with at least three columns with 3d coordinates (i.e X,Y,Z)}

\item{keep_original}{logical (Default = FALSE). Should the original branching order and axis be kept ? Otherwise, it is re-estimated.}
}
\description{
Read an aRchi file
}
\examples{
# Create an empty object of class aRchi
 empty_aRchi=aRchi()
 # Write the aRchi object
write_aRchi(empty_aRchi,file="my_empty_aRchi_file.aRchi")
read_aRchi(file = "my_empty_aRchi_file.aRchi")

}
\seealso{
[write_aRchi()]; [read_aRchi()]
}
