% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_A0.R
\docType{methods}
\name{Compute_A0}
\alias{Compute_A0}
\title{Find the principal axis of a tree}
\usage{
Compute_A0(aRchi, plotresult = FALSE)
}
\arguments{
\item{aRchi}{an object of class aRchi with at least a QSM and the Paths table}

\item{plotresult}{logical (default = FALSE). Show the results in a 3d plot if TRUE}
}
\value{
The aRchi object with the QSM having a new column A0.
}
\description{
Find the principal axis of a tree (i.e A0) and add a column to the QSM of an aRchi object. This is alternative method to the default branch order proposed in a QSM for the principal axis only.
}
\details{
The method used to find the principal axis consist in finding the highest vertical path with consecutive segments of similar diameters and orientations. An index called A0 of the probability of being the principal axis is thus computed for each path of the tree and the path with the highest value is considered as the principal axis.
This A0 index is calculated as follow: A0=(1-ΔAng)+(1-ΔDiam)+H_P+¯Ang where ΔAng and ΔDiam are the mean relative differences of zenithal angles and diameters (last cylinder diameter for parent and first for daughter) between two consecutive segments (i.e the mother and the daughter of the same path), respectively. H_P is the relative height of the path, corresponding to the ratio between the height of the highest segment of the path and total tree height; and ¯Ang is the mean relative zenithal angle of all segments in a path, which varies between 0 (perfectly horizontal) and 1 (perfectly vertical).

A0 ranges between 0 and 4 with 0 indicating a path with a low probability of being the principal axis and 4 indicating a high probability of being the principal axis, i.e. the highest vertical path with consecutive segments of similar diameters and orientations.The path with the maximum A0 value was selected as the principal axis.

The new column A0 of the QSM slot take the value 2 if the cylinder is part of the principal axis or 1 if not.

#'@references

	Martin-Ducup, O. et al. Terrestrial laser scanning reveals convergence of tree architecture with increasingly dominant crown canopy position. Functional Ecology (2020).
}
\examples{
# Read an aRchi file with at least the QSM and the paths table
file=system.file("extdata","Tree_1_aRchi.aRchi",package = "aRchi")
Tree1_aRchi=read_aRchi(file)

Tree1_aRchi=Compute_A0(Tree1_aRchi,plotresult=TRUE)

}
\seealso{
[DAI())] to compute the dominance of a principal axis index that uses the A0 index.
}
