% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prod_3d.R
\name{cross_prod_3d}
\alias{cross_prod_3d}
\title{Internal function that performs a cross product for the dist2line function}
\usage{
cross_prod_3d(v1, v2)
}
\arguments{
\item{v1}{vector 1}

\item{v2}{vector 2}
}
\description{
Internal function that performs a cross product for the dist2line function
}
\keyword{internal}
