% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2line.R
\name{dist2line}
\alias{dist2line}
\title{Internal function to compute the distance of point b to segment ac}
\usage{
dist2line(a, b, c)
}
\arguments{
\item{a}{a vector containing the xyz coordinates of a point}

\item{b}{a vector containing the xyz coordinates of a point}

\item{c}{a vector containing the xyz coordinates of a point}
}
\value{
the distance of point b to segment ac
}
\description{
Internal function to compute the distance of point b to segment ac
}
\keyword{internal}
