% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_QSM.R
\docType{methods}
\name{get_QSM}
\alias{get_QSM}
\alias{get_QSM,aRchi-method}
\title{Get the QSM from an object of class aRchi}
\usage{
get_QSM(aRchi)

\S4method{get_QSM}{aRchi}(aRchi)
}
\arguments{
\item{aRchi}{The object of class aRchi}
}
\description{
Get the QSM from an object of class aRchi
}
\examples{
# Read an aRchi file with a QSM and paths tables.
file=system.file("extdata","Tree_1_aRchi.aRchi",package = "aRchi")
Tree1_aRchi=read_aRchi(file)

# show the QSM data.table
get_QSM(Tree1_aRchi)

}
\seealso{
\code{\link{get_pointcloud}}; \code{\link{get_paths}}; \code{\link{get_nodes}}
}
