% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSM2mesh.R
\name{QSM2mesh}
\alias{QSM2mesh}
\title{Transform a QSM into a triangular or quadrangular mesh}
\usage{
QSM2mesh(qsm, tmesh = FALSE, sides = 16)
}
\arguments{
\item{qsm}{a qsm.}

\item{tmesh}{logical. If TRUE a triangular mesh is returned, otherwise a quadrangular mesh is returned.}

\item{sides}{integer. The number of faces that compose a cylinder.}
}
\value{
a 3D mesh.
}
\description{
Transform a QSM into a triangular or quadrangular mesh
}
\keyword{internal}
