% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aRchi.R
\docType{methods}
\name{read_aRchi}
\alias{read_aRchi}
\title{Read an aRchi file}
\usage{
read_aRchi(file)
}
\arguments{
\item{file}{The directory to the `.aRchi` file.}
}
\description{
Read an aRchi file
}
\examples{
# Create an empty object of class aRchi
 empty_aRchi=aRchi()
 # Write the aRchi object
write_aRchi(empty_aRchi,file="my_empty_aRchi_file.aRchi")
read_aRchi(file = "my_empty_aRchi_file.aRchi")

}
\seealso{
[write_aRchi()]
}
