% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_QSM.R
\name{read_QSM}
\alias{read_QSM}
\title{Read a QSM}
\usage{
read_QSM(file, model)
}
\arguments{
\item{file}{The directory to the QSM file path}

\item{model}{`treeQSM`, `simpletree`, `simpleforest` or `pypetree` depending on the algorithm used to generate the QSM}
}
\value{
a data.table that can be used to build an aRchi object (see function [build_aRchi())])
}
\description{
Read a QSM file generated with treeQSM, simpletree, simpleforest or pypetree.
}
\examples{
file=system.file("extdata","Tree_1_TreeQSM.txt",package = "aRchi")
QSM=read_QSM(file,model="treeQSM")
}
\seealso{
[aRchi()] the aRchi class;[build_aRchi())] to build an object of class `aRchi`; [skeletonize())]; to make a QSM from a point cloud in an aRchi object.
}
