% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_aRchi.R
\docType{methods}
\name{write_aRchi}
\alias{write_aRchi}
\alias{write_aRchi,aRchi,character-method}
\title{Write an aRchi file}
\usage{
write_aRchi(aRchi, file)

\S4method{write_aRchi}{aRchi,character}(aRchi, file)
}
\arguments{
\item{aRchi}{the object of class aRchi to write}

\item{file}{The directory where to write the .aRchi file.}
}
\description{
write an aRchi file
}
\seealso{
\code{\link{read_aRchi}}
}
