% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRchiClass.R
\docType{class}
\name{aRchi-class}
\alias{aRchi-class}
\alias{aRchi}
\title{aRchi}
\value{
An object of class aRchi
}
\description{
Class containing files to compute and display in three dimensions tree architectural metrics at different level of organization
}
\section{Fields}{

\describe{
\item{\code{QSM}}{a data.table containing QSM information according to read_QSM function format}

\item{\code{pointcloud}}{a data.table containing the point cloud used to generated the QSM}

\item{\code{Paths}}{a data.table of Paths according to Make_Path function (see \code{\link{Make_Path}})}

\item{\code{Nodes}}{Metrics computed at the node scale (see \code{\link{Make_Node}}}

\item{\code{operations}}{Record all the operations realized on the object.}
}}

