% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodes.R
\docType{methods}
\name{get_nodes}
\alias{get_nodes}
\alias{get_nodes,aRchi-method}
\title{Get the nodes from an object of class aRchi}
\usage{
get_nodes(aRchi)

\S4method{get_nodes}{aRchi}(aRchi)
}
\arguments{
\item{aRchi}{The object of class aRchi}
}
\description{
Get the nodes from an object of class aRchi
}
\examples{
# Read an aRchi file with a QSM and paths tables.
file=system.file("extdata","Tree_1_aRchi.aRchi",package = "aRchi")
Tree1_aRchi=read_aRchi(file)

# get the nodes (a list of two data.table)
get_nodes(Tree1_aRchi)

}
\seealso{
\code{\link{get_QSM}}; \code{\link{get_pointcloud}}; \code{\link{get_paths}}
}
