% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\docType{methods}
\name{get_paths}
\alias{get_paths}
\title{Get the paths from an object of class aRchi}
\usage{
get_paths(aRchi)
}
\arguments{
\item{aRchi}{The object of class aRchi}
}
\description{
Get the paths from an object of class aRchi
}
\examples{
# Read an aRchi file with a QSM and paths tables.
file=system.file("extdata","Tree_1_aRchi.aRchi",package = "aRchi")
Tree1_aRchi=read_aRchi(file)

# get the paths (a data.table)
get_paths(Tree1_aRchi)

}
\seealso{
[get_QSM()]; [get_pointcloud()]; [get_nodes()];
}
