% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pointcloud.R
\docType{methods}
\name{get_pointcloud}
\alias{get_pointcloud}
\title{Get the pointcloud from an object of class aRchi}
\usage{
get_pointcloud(aRchi)
}
\arguments{
\item{aRchi}{The object of class aRchi}
}
\description{
Get the pointcloud from an object of class aRchi
}
\examples{
# Read an aRchi file with a QSM and paths tables.
file=system.file("extdata","Tree_1_aRchi.aRchi",package = "aRchi")
Tree1_aRchi=read_aRchi(file)

# get the pointcloud
get_pointcloud(Tree1_aRchi)

}
\seealso{
[get_QSM()]; [get_paths()]; [get_nodes()];
}
