% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aRchi.R
\docType{methods}
\name{read_aRchi}
\alias{read_aRchi}
\alias{read_aRchi,character-method}
\title{Read an aRchi file}
\usage{
read_aRchi(file)

\S4method{read_aRchi}{character}(file)
}
\arguments{
\item{file}{The directory to the .aRchi file.}
}
\description{
Read an aRchi file
}
\seealso{
\code{\link{write_aRchi}}
}
