% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_aRchi.R
\docType{methods}
\name{write_aRchi}
\alias{write_aRchi}
\title{Write an aRchi file}
\usage{
write_aRchi(x, file)
}
\arguments{
\item{file}{The directory where to write the `.aRchi` file.}

\item{aRchi}{the object of class aRchi to write}
}
\description{
write an aRchi file
}
\examples{
# Create an empty object of class aRchi
 empty_aRchi=aRchi()
 # Write the aRchi object
write_aRchi(empty_aRchi,file="my_empty_aRchi_file.aRchi")
read_aRchi(file = "my_empty_aRchi_file.aRchi")
}
\seealso{
[read_aRchi()]
}
